+++ src/com/kaylerrenslow/armaDialogCreator/gui/uicanvas/Region.java
@@ -1,6 +1,7 @@
 package com.kaylerrenslow.armaDialogCreator.gui.uicanvas;
 
 import javafx.scene.canvas.GraphicsContext;
+import javafx.scene.paint.Color;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -109,6 +110,47 @@ public interface Region {
 	}
 
 	/**
+	 Paint a checkerboard with the given {@link GraphicsContext} and coordinates
+
+	 @param gc context
+	 @param x x pos
+	 @param y y pos
+	 @param w width
+	 @param h height
+	 @param color1 one of the colors to use
+	 @param color2 other color to use
+	 @param numBoxes number of boxes across horizontal and vertical
+	 */
+	static void paintCheckerboard(@NotNull GraphicsContext gc, int x, int y, int w, int h,
+								  @NotNull Color color1, @NotNull Color color2, int numBoxes) {
+		final int boxWidth = w / numBoxes;
+		final int boxHeight = h / numBoxes;
+		final int remainderWidth = w - numBoxes * boxWidth;
+		final int remainderHeight = h - numBoxes * boxHeight;
+
+		gc.save();
+
+		for (int row = 0; row < numBoxes || (row == numBoxes && remainderHeight > 0); row++) {
+			//doing <= to make sure that the full height and width is painted
+			//since the aspect ratio of the checkerboard may not be equal to aspect ratio of the given width and height
+
+			int yy = y + row * boxHeight;
+			for (int box = 0; box < numBoxes || (box == numBoxes && remainderWidth > 0); box++) {
+				gc.setStroke((box + row) % 2 == 0 ? color1 : color2);
+				int xx = x + box * boxWidth;
+				fillRectangle(
+						gc, xx, yy,
+						xx + (box == numBoxes ? remainderWidth : boxWidth),
+						yy + (row == numBoxes ? remainderHeight : boxHeight)
+				);
+			}
+		}
+
+		gc.restore();
+	}
+
+
+	/**
 	 Translate the region's x and y coordinates relative to the given point. Even if this region isn't allowed to move, this method will work.
 
 	 @param dx change in x
@@ -142,6 +184,7 @@ public interface Region {
 		}
 		return false;
 	}
+
 	/**
 	 Gets the edge(s) that the given point is closest to.