+++ CHANGES.md
@@ -1,11 +1,17 @@
 **Added:**
-* BasicTextRenderer can now do multiple lines
+* Edit control renderer
+* Edit control data creator
 
 **Changed:**
-* 
+* static can now support multi line
+* in canvas editor: when a control is no longer visible, it is removed from the selection
+* removed auto submit from color array editor popup
+* color array editor  popup now has a transparency grid to help show there is transparency
 
 **Fixed:**
-* 
+* sometimes the option value editor was becoming too big and wasn't scaling according to the editor container
+* non multi lined text wasn't rendering in correct spot upon renderer init
+* double clicking a control that was selected but the mouse over control wasn't matched with the selected control resulted in the editor popup not showing
 
 **Notes:**
 * May 31: we should have default value providers in one xml file for every control.